Implémentation des vues avec Blade et refactoring du code

+++ app/Services/AuthService.php
@@ -7,7 +7,9 @@ use App\Models\VerificationCode;
 use App\Repositories\AuthRepository;
 use Illuminate\Support\Facades\Mail;
 use App\Mail\TwoFactorCodeMail;
+use App\Mail\RestoreAccountCode;
 use Illuminate\Validation\ValidationException;
+use Illuminate\Support\Facades\Hash;
 
 class AuthService 
 {
@@ -19,28 +21,34 @@ class AuthService
         $this->authRepository = $authRepository;
     }
 
+    public function createUser(array $data) 
+    {
+        // Hash obligatoire
+        $data['password'] = Hash::make($data['password']);
+        
+        $user = $this->authRepository->create($data);
+
+        return [
+            'success' => true,
+            'message' => 'Compte créé avec succès ! Vérifiez votre boîte mail pour activer votre compte',
+            'data' => $user,
+            'status' => 200
+        ];
+    }
+
     // -@- Connexion d'un utilisateur
     public function login(array $credentials) 
     {
         // -@- Etape 1 : On cherche l'utilisateur actif ou supprimé.
 
-        $user = User::withTrashed()->where('email', $credentials['email'])->first();
+        $user = $this->authRepository->findUserByEmail($credentials['email']);
 
-        if (!$user) {
+        if (!$user || !$this->authRepository->verifyPassword($user, $credentials['password'])) {
             return [
                 'success' => false, 
                 'message' => 'Identifiants incorrects.', 
-                'status' => 401
+                'status' => 401                                                                                                                                                                                                                                                                                                                  
             ];
-        }
-
-        if (!$this->authRepository->verifyPassword($user, $credentials['password'])) {
-           return [
-                'success' => false, 
-                'message' => 'Identifiants incorrects.', 
-                'status' => 401
-            ];
-        
         }
 
         if ($user->trashed()) {
@@ -54,51 +62,25 @@ class AuthService
 
         // -@- Etape 2FA : Génération du code.
         $code = str_pad(random_int(0, 999999), 6, '0', STR_PAD_LEFT);
-
-        VerificationCode::create([
-            'user_id' => $user->id,
-            'code' => $code,
-            'expires_at' => now()->addMinutes(10)
-        ]);
+        $this->authRepository->createVerificationCode($user->id, $code);
 
         // -@- Envoi de l'email
         Mail::to($user->email)->send(new TwoFactorCodeMail($code));
 
-        // -@- Reception du code par mail via mailtrap. 
-        // -@- Simuler l'envoi (Envoi dans la réponse pour les tests)
         return [
             'success' => true,
             'status' => '2FA_REQUIRED',
             'message' => 'Un code de vérification a été envoyé à votre adresse mail.',
             'user_id' => $user->id
-            // 'message' => 'Veuillez entrer le code envoyé par email',
-            // 'temp_user_id' => $user->id,
             // 'dev_debug_code' => $code 
         ];
-
-        // Simulation de l'envoi du SMS
-        // if(config('services.sms.provider') === 'log') {
-        //     \Illuminate\Support\Facades\Log::info("SMS envoyé à {$user->phone} : Votre code est $code ");
-        // }
-
-        // Suppression des anciens tokens pour n'en avoir qu'un actif
-       //  $user->tokens()->delete();
-
-        // Génération du token Sanctum
-        // return [
-        //     'user' => $user,
-        //     'token' => $user->createToken('api-token')->plainTextToken
-        // ];
     }
 
     // -@- Vérification 2FA
-    public function verify2FACode(string $userId, string $code) 
+    public function verify2fa(string $userId, string $code) 
     {
-        // -@- 1. Chercher le dernier code valide pour cet UUID
-        $verification = \App\Models\VerificationCode::where('user_id',$userId)
-        ->latest()
-        ->first();
-
+        $verification = $this->authRepository->findVerificationCode($userId);
+      
         // -@- 2. Vérifier s'il existe 
         if(!$verification){
             return [
@@ -161,9 +143,102 @@ class AuthService
         ];         
     }
 
-    // 8. -@- Déconnexion de l'utilisateur
-    public function logout($user) 
+    public function getAuthenticatedUser($request)
     {
-        return $user->currentAccessToken()->delete();
+        $user = $this->authRepository->getAuthenticatedUser($request);
+
+        if(!$user) {
+            return [
+                'success' => false,
+                'message' => 'Non authentifié',
+                'status' => 401
+            ];
+        }
+
+        return [
+            'success' => true,
+            'message' => 'Utilisateur connecté récupéré avec succès',
+            'data' => $user,
+            'status' => 200
+        ];
+    }
+
+    // -@- 
+    public  function requestRestoration(string $email) 
+    {
+        $user = $this->authRepository->findUserTrashed($email);
+
+        if(!$user) {
+            return [
+                'success' => false,
+                'message' => 'Aucun compte supprimé trouvé',
+                'status' => 404
+            ];
+        }
+
+        $code = str_pad(random_int(0, 999999), 6, '0', STR_PAD_LEFT);
+        
+        $this->authRepository->createVerificationCode($user->id, $code);
+
+        Mail::to($user->email)->send(new RestoreAccountCode($code));
+
+        return [
+            'success' => true,
+            'message' => 'Un code de vérification a été envoyé sur votre boîte mail.',
+            'status' => 200
+        ];
+    }
+
+    public function confirmRestoration(string $email, string $code) 
+    {
+        $user = $this->authRepository->findUserTrashed($email);
+
+        if(!$user) {
+            return [
+                'success' => false,
+                'message' => 'Utilisateur introuvable',
+                'status' => 404
+            ];
+        }
+
+        $verification = $this->authRepository->findRestaurationCode($user->id, $code);
+
+        if(!$verification) {
+            return [
+                'success' => false,
+                'message' => 'code invalide ou expiré',
+                'status' => 422
+            ];
+        }
+
+        $this->authRepository->restore($user);
+        $verification->delete();
+
+        $user->notify(new \App\Notifications\AccountRestoreNotification());
+
+        return [
+            'success' => true,
+            'message' => 'Compte restauré avec succès',
+            'status' => 200,
+            'data' => $user
+        ];
+    }
+
+    // 8. -@- Déconnexion de l'utilisateur
+    public function logout($user, bool $api = true) 
+    {
+        if($api) {
+            return $user->currentAccessToken()->delete();
+        }
+
+       /** @var \Illuminate\Contracts\Auth\StatefulGuard $guard */
+        $guard = auth();
+        $guard->logout(); 
+        
+        session()->invalidate();
+        session()->regenerateToken();
+
+        return true;
+
     }
 }
